move e2e tests to a new package
