+++ lage.config.js
@@ -4,9 +4,10 @@
 module.exports = {
   pipeline: {
     build: ["^build"],
-    test: ["@lage-run/monorepo-fixture#build"],
+    test: [],
     lint: [],
     start: [],
+    "@lage-run/e2e-tests#test": ["^build"],
 
     // TODO: a temporary hack to allow both of these projects to run build with lage
     "@lage-run/docs-beta#build": ["@lage/docs#build"]
