+++ packages/e2e-tests/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require("monorepo-scripts/config/jest.config.js");
