+++ packages/e2e-tests/src/info.test-disabled.ts
@@ -1,4 +1,4 @@
-import { Monorepo } from "../mock/monorepo";
+import { Monorepo } from "./mock/monorepo";
 
 describe("info test", () => {
   it("with json reporter", () => {
