+++ packages/e2e-tests/src/mock/monorepo.ts
@@ -174,6 +174,6 @@ export class Monorepo {
   }
 
   cleanup() {
-    fs.rmdirSync(this.root, { recursive: true });
+    fs.rmSync(this.root, { recursive: true });
   }
 }
