+++ packages/lage/package.json
@@ -59,7 +59,7 @@
     "@types/git-url-parse": "9.0.1",
     "@types/ioredis": "4.28.10",
     "@types/jest": "^27.0.1",
-    "@types/node": "12.20.55",
+    "@types/node": "^14.0.0",
     "@types/npmlog": "4.1.4",
     "@types/p-queue": "3.2.1",
     "@types/yargs-parser": "15.0.0",
