+++ packages/logger/package.json
@@ -15,7 +15,7 @@
   },
   "devDependencies": {
     "@types/jest": "^27.0.1",
-    "@types/node": "12.20.55",
+    "@types/node": "^14.0.0",
     "jest": "^27.2.0",
     "ts-jest": "^27.0.5",
     "typescript": "^4.4.4"
