+++ packages/reporters/package.json
@@ -21,7 +21,7 @@
   },
   "devDependencies": {
     "@types/jest": "^27.0.1",
-    "@types/node": "12.20.55",
+    "@types/node": "^14.0.0",
     "jest": "^27.2.0",
     "memory-streams": "^0.1.3",
     "ts-jest": "^27.0.5",
