+++ packages/scheduler/tests/SimpleScheduler.test.ts
@@ -290,62 +290,62 @@ describe("SimpleScheduler", () => {
     const summary = await schedulerPromise;
 
     expect(dropTiming(summary)).toMatchInlineSnapshot(`
-      Object {
-        "error": undefined,
-        "results": "aborted",
-        "targetRunByStatus": Object {
-          "aborted": Array [],
-          "failed": Array [],
-          "pending": Array [
-            "a#build",
-            "b#build",
-            "c#build",
-            "d#build",
-            "e#build",
-            "f#build",
-            "g#build",
-          ],
-          "running": Array [],
-          "skipped": Array [],
-          "success": Array [
-            "__start",
-          ],
-        },
-        "targetRuns": Map {
-          "__start" => Object {
-            "status": "success",
-            "target": "__start",
-          },
-          "a#build" => Object {
-            "status": "pending",
-            "target": "a#build",
-          },
-          "b#build" => Object {
-            "status": "pending",
-            "target": "b#build",
-          },
-          "c#build" => Object {
-            "status": "pending",
-            "target": "c#build",
-          },
-          "d#build" => Object {
-            "status": "pending",
-            "target": "d#build",
-          },
-          "e#build" => Object {
-            "status": "pending",
-            "target": "e#build",
-          },
-          "f#build" => Object {
-            "status": "pending",
-            "target": "f#build",
-          },
-          "g#build" => Object {
-            "status": "pending",
-            "target": "g#build",
-          },
-        },
-      }
-    `);
+Object {
+  "error": undefined,
+  "results": "failed",
+  "targetRunByStatus": Object {
+    "aborted": Array [],
+    "failed": Array [],
+    "pending": Array [
+      "a#build",
+      "b#build",
+      "c#build",
+      "d#build",
+      "e#build",
+      "f#build",
+      "g#build",
+    ],
+    "running": Array [],
+    "skipped": Array [],
+    "success": Array [
+      "__start",
+    ],
+  },
+  "targetRuns": Map {
+    "__start" => Object {
+      "status": "success",
+      "target": "__start",
+    },
+    "a#build" => Object {
+      "status": "pending",
+      "target": "a#build",
+    },
+    "b#build" => Object {
+      "status": "pending",
+      "target": "b#build",
+    },
+    "c#build" => Object {
+      "status": "pending",
+      "target": "c#build",
+    },
+    "d#build" => Object {
+      "status": "pending",
+      "target": "d#build",
+    },
+    "e#build" => Object {
+      "status": "pending",
+      "target": "e#build",
+    },
+    "f#build" => Object {
+      "status": "pending",
+      "target": "f#build",
+    },
+    "g#build" => Object {
+      "status": "pending",
+      "target": "g#build",
+    },
+  },
+}
+`);
   });
 });
